refactor: improve streamTranslate()
