--- Easydict/Swift/Feature/HTTPServer/QueryService+Extension/BaseOpenAIService+Stream.swift
@@ -1,9 +1,0 @@
-//
-//  BaseOpenAIService+Stream.swift
-//  Easydict
-//
-//  Created by tisfeng on 2024/7/29.
-//  Copyright © 2024 izual. All rights reserved.
-//
-
-import Foundation
